% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align_forecasts_one_temporal_resolution}
\alias{align_forecasts_one_temporal_resolution}
\title{Internal function that add reference dates and relative horizons
to a dataframe of forecasts. This function requires that the forecasts
data frame contains only forecasts one temporal_resolution, i.e., only
at the daily resolution or the weekly resolution.}
\usage{
align_forecasts_one_temporal_resolution(
  forecasts,
  reference_dates,
  reference_weekday,
  reference_windows,
  drop_nonpos_relative_horizons
)
}
\arguments{
\item{forecasts}{dataframe in format returned by \code{load_forecasts}}

\item{reference_dates}{a vector of reference dates for forecasts, grouping
forecasts that were made during the same week. May be a character vector in
the format "YYYY-MM-DD" or a vector of Dates. For example, if a forecast of
daily hospitalizations is issued with a forecast date that is Saturday,
2021-09-04, and we want to align analyses around a Monday, the reference
date would be Monday, 2021-09-06. This can be accomplished more easily by
providing the \code{reference_weekday} argument below.}

\item{reference_weekday}{capitalized weekday, e.g. "Monday". The day of week
to use for reference dates, if the \code{reference_dates} argument is missing or
\code{NULL}}

\item{reference_windows}{an integer vector of offset values indicating the
set of forecast dates should be assigned to a particular reference date,
in units of number of days. If the \code{reference_weekday} is "Monday", the
function defaults to \code{-6:0}, so that all forecasts with a \code{forecast_date}
in the week leading up to a particular Monday will be assigned a reference
date of that Monday. By default, if the \code{reference_weekday} is "Saturday",
the function defaults to \code{-4:2}. This is appropriate for forecasts of
weekly targets, in which case we might like to assign a forecast issued on
a Monday to have a reference date of the previous Saturday.}

\item{drop_nonpos_relative_horizons}{boolean indicating whether forecasts that have a
non-positive horizon relative to the reference date should be dropped.
Defaults to \code{TRUE}}
}
\value{
forecast dataframe augmented by columns reference_date and
relative_horizon
}
\description{
Internal function that add reference dates and relative horizons
to a dataframe of forecasts. This function requires that the forecasts
data frame contains only forecasts one temporal_resolution, i.e., only
at the daily resolution or the weekly resolution.
}
