% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_covid_hub_forecasts.R
\name{as_covid_hub_forecasts}
\alias{as_covid_hub_forecasts}
\title{Reformat model outputs stored as a \code{model_output_tbl} class (or similar) to
that of a \code{data.frame} formatted according to standards of the COVID-19
Forecasting Hub which can be processed by functions from the \code{covidHubUtils}
package such as \code{score_forecasts()} or \code{plot_forecasts()}. The supplied
\code{model_output_tbl} should have columns defining properties akin to
reference dates, locations, horizons, and targets.}
\usage{
as_covid_hub_forecasts(
  model_outputs,
  model_id_col = "model_id",
  reference_date_col = "forecast_date",
  location_col = "location",
  horizon_col = "horizon",
  target_col = "target",
  output_type_col = "output_type",
  output_type_id_col = "output_type_id",
  value_col = "value",
  temp_res_col = "temporal_resolution",
  target_end_date_col = "target_end_date"
)
}
\arguments{
\item{model_outputs}{an object of class \code{model_output_tbl} with component
model outputs (e.g., predictions). Should have columns containing the
following information: model name, reference date or target end date,
location, horizon, target, temporal resolution*, output type, output
type id, and value. Note that the temporal resolution may be included
in the target column.}

\item{model_id_col}{\code{character} string of the name of the column
containing the model name(s) for the forecasts. Defaults to "model_id".
Should be set to NULL if no such column exists, in which case a model_id
column will be created populated with the value "model_id".}

\item{reference_date_col}{\code{character} string of the name of the column
containing the reference dates for the forecasts. Defaults to
"forecast_date". Should be set to NULL if no such column exists, in which
case the column will be created using the following information:
horizon, target end date, and temporal resolution.}

\item{location_col}{\code{character} string of the name of the column
containing the locations for the forecasts. Defaults to "location".}

\item{horizon_col}{\code{character} string of the name of the column
containing the horizons for the forecasts. Defaults to "horizon".}

\item{target_col}{\code{character} string of the name of the column
containing the targets for the forecasts. Defaults to "target". If
\code{temp_res_col} is NULL, the target column in \code{model_outputs} is assumed
to contain targets of the form "temporal_resolution target" or
"temporal_resolution ahead target", such as "wk ahead inc flu hosp"
"wk inc flu hosp".}

\item{output_type_col}{\code{character} string of the name of the column
containing the output types for the forecasts. Defaults to "output_type".}

\item{output_type_id_col}{\code{character} string of the name of the column
containing the output type ids for the forecasts. Defaults to
"output_type_id".}

\item{value_col}{\code{character} string of the name of the column
containing the values for the forecasts. Defaults to "value".}

\item{temp_res_col}{\code{character} string of the name of the column
containing the temporal resolutions for the forecasts. Defaults to
"temporal_resolution". Should be set to NULL if no such column exists,
in which case the column will be created from the existing target column.}

\item{target_end_date_col}{\code{character} string of the name of the column
containing the target end dates for the forecasts. Defaults to
"target_end_date". Should be set to NULL if no such column exists, in
which case the column will be created using the following information:
horizon, forecast date, and temporal resolution.}
}
\value{
a \code{data.frame} of reformatted model outputs that may be fed into
any of the \code{covidHubUtils} functions with 10 total columns: model,
forecast_date, location, horizon, temporal_resolution, target_variable,
target_end_date, type, quantile, value. Other columns are removed.
}
\description{
Reformat model outputs stored as a \code{model_output_tbl} class (or similar) to
that of a \code{data.frame} formatted according to standards of the COVID-19
Forecasting Hub which can be processed by functions from the \code{covidHubUtils}
package such as \code{score_forecasts()} or \code{plot_forecasts()}. The supplied
\code{model_output_tbl} should have columns defining properties akin to
reference dates, locations, horizons, and targets.
}
\examples{
forecasts <- load_forecasts(
  models = c("COVIDhub-ensemble", "UMass-MechBayes"),
  dates = "2020-12-14",
  date_window_size = 7,
  locations = c("US"),
  targets = paste(1:4, "wk ahead inc death"),
  source = "zoltar"
) 
altered_forecasts <- forecasts |> # Alter forecasts to not be CovidHub format
  mutate(target_variable = "wk ahead inc death", horizon=as.numeric(horizon)) |>
  select(-temporal_resolution)
formatted_forecasts <- as_covid_hub_forecasts(
   altered_forecasts, 
   target_col="target_variable", 
   temp_res_col=NULL 
) |>
mutate(horizon=as.character(horizon))
expect_equal(formatted_forecasts, dplyr::select(forecasts, model:value)) 
}
