% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cramers_dist_equal_space.R
\name{calc_cramers_dist_equal_space}
\alias{calc_cramers_dist_equal_space}
\title{Approximate the Cramer’s distance between a pair of distributions
F and G that are represented by a collection of equally-spaced quantiles.}
\usage{
calc_cramers_dist_equal_space(q_F, tau_F, q_G, tau_G, approx_rule)
}
\arguments{
\item{q_F}{vector containing the quantiles of F}

\item{tau_F}{vector containing the probability levels corresponding to
the quantiles of F.}

\item{q_G}{vector containing the quantiles of G}

\item{tau_G}{vector containing the probability levels corresponding to
the quantiles of G.}

\item{approx_rule}{string specifying which formula to use
for approximation. Valid rules are "approximation1" and
"approximation2".
See Details for more information.}
}
\value{
a single value of approximated pairwise Cramér distance
between q_F and q_G
}
\description{
Approximate the Cramer’s distance between a pair of distributions
F and G that are represented by a collection of equally-spaced quantiles.
}
\details{
This function requires the two vectors of quantiles to be
of equal length. The approximation methods are formulated based on
two collections of quantiles corresponding to equally-spaced
probability levels. The approximation formula for "approximation1" is
\deqn{ \text{CD}(F,G) \approx \left\{\frac{1}{K(K+1)}\sum^{2K-1}_{i=1}b_i(b_i+1)(q_{i+1}-q_i)\right }{ (non-Latex version) }
and the approximation formula for "approximation2" is
\deqn{ \text{CD}(F,G) \approx \left\{\frac{1}{(K+1)^2}\sum^{2K-1}_{i=1}b_i^2(q_{i+1}-q_i)\right. }{ (non-Latex version) }
where \eqn{q_i} is an element in a vector of an ordered pooled quantiles
of \code{q_F} and \code{q_G} and \eqn{b_i} is an element of a vector of the absolute
values of cumulative sums of \eqn{\mathbf{a}}, whose element is 1 if
\eqn{q_i} is a quantile of F or -1 if \eqn{q_i} is a quantile of G.
}
