% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cramers_dist_unequal_space.R
\name{calc_cramers_dist_unequal_space}
\alias{calc_cramers_dist_unequal_space}
\title{Approximate the Cramer’s distance between a pair of distributions
F and G that are represented by a collection of unequally-spaced quantiles.}
\usage{
calc_cramers_dist_unequal_space(q_F, tau_F, q_G, tau_G, approx_rule)
}
\arguments{
\item{q_F}{vector containing the quantiles of F}

\item{tau_F}{vector containing the probability levels corresponding to
the quantiles of F.}

\item{q_G}{vector containing the quantiles of G}

\item{tau_G}{vector containing the probability levels corresponding to
the quantiles of G.}

\item{approx_rule}{string specifying which formula to use
for approximation. Valid rules are "left_sided_riemann" and
"trapezoid_riemann". See Details for more information.}
}
\value{
a single value of approximated pairwise Cramér distance
between q_F and q_G
}
\description{
Approximate the Cramer’s distance between a pair of distributions
F and G that are represented by a collection of unequally-spaced quantiles.
}
\details{
This function accommodates cases when the lengths of
\code{q_F} and \code{q_G} are not equal and when \code{tau_F} and \code{tau_G} are not equal.
A Riemann sum is used to approximate a pairwise Cramér distance.
The approximation formula for "left_sided_riemann" is
\deqn{ \text{CD}(F,G) \approx \left\{\sum^{2K-1}_{j=1}(\tau^F_j-\tau^G_j)^2(q_{i+1}-q_i)\right }{ (non-Latex version) }
and the approximation formula for "trapezoid_riemann" is
\deqn{ \text{CD}(F,G) \approx \left\{\frac{1}{(K+1)^2}\sum^{2K-1}_{i=1}\frac{(\tau^F_j-\tau^G_j)^2+(\tau^F_{j+1}-\tau^G_{j+1})^2}{2}(q_{i+1}-q_i)\right. }{ (non-Latex version) }
where \eqn{q_i} is an element in a vector of an ordered pooled quantiles
of \code{q_F} and \code{q_G} and \eqn{\tau^F_j} and \eqn{\tau^G_j} are defined as
the probability level of a quantile in \code{q_F} when \eqn{q_i} comes from \eqn{F} and
the probability level of a quantile in \code{q_G} when \eqn{q_i} comes from \eqn{G},
respectively.
}
