% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zoltar_truth.R
\name{configure_zoltar_truth}
\alias{configure_zoltar_truth}
\title{Generate the most up to date truth data for zoltar for a specified target
It only includes national and state-level truth data.}
\usage{
configure_zoltar_truth(target)
}
\arguments{
\item{target}{string specifying target.
Currently only support "Cumulative Deaths" and "Incident Deaths"}
}
\value{
data frame with timezero, unit, target and value
}
\description{
Generate the most up to date truth data for zoltar for a specified target
It only includes national and state-level truth data.
}
