% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{date_to_datetime}
\alias{date_to_datetime}
\title{Create a datetime by appending time and/or timezone
to the given date string, if necessary}
\usage{
date_to_datetime(date, hub = c("US", "ECDC"))
}
\arguments{
\item{date}{date in character.
It could be a date, or a date with time, or a date with time and timezone.}

\item{hub}{character vector, where the first element indicates the hub
to set default timezone. Possible options are "US" and "ECDC".}
}
\value{
datetime characters with date, time and timezone
}
\description{
Create a datetime by appending time and/or timezone
to the given date string, if necessary
}
