% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{date_to_week_end_date}
\alias{date_to_week_end_date}
\title{Calculate the date of the Saturday that ends the
MMWR epidemic week for the given date,offset a specified number
of epidemic week from a specified date}
\usage{
date_to_week_end_date(forecast_date, week_offset = 0)
}
\arguments{
\item{forecast_date}{vector of dates as Date objects}

\item{week_offset}{vector of week offsets.  must be either length 1 or the same
length as forecast_date}
}
\value{
character vector of dates in 'yyyy-mm-dd' format
}
\description{
Calculate the date of the Saturday that ends the
MMWR epidemic week for the given date,offset a specified number
of epidemic week from a specified date
}
