% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{download_raw_nytimes}
\alias{download_raw_nytimes}
\title{Download raw truth data from NYTimes and write to CSV files}
\usage{
download_raw_nytimes(save_location = "./data-truth/nytimes/raw/")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./data-truth/nytimes/raw/"}
}
\value{
data frame of national, state and county level raw truth data
}
\description{
Download raw truth data from NYTimes and write to CSV files
}
