% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{download_raw_usafacts}
\alias{download_raw_usafacts}
\title{Download raw truth data from USAFacts}
\usage{
download_raw_usafacts(save_location = "./data-truth/usafacts/raw/")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./data-truth/usafacts/raw/"}
}
\value{
data frame of cases and deaths raw truth data
}
\description{
Download raw truth data from USAFacts
}
