% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_model_abbr.R
\name{get_all_model_abbr}
\alias{get_all_model_abbr}
\title{Get all valid model_abbrs from remote hub repo}
\usage{
get_all_model_abbr()
}
\value{
a list of valid model names
}
\description{
Get all valid model_abbrs from remote hub repo
}
