% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_latest_forecasts_repo.R
\name{get_forecast_file_path}
\alias{get_forecast_file_path}
\title{Generate forecast file paths with given data-processed folder path for selected models}
\usage{
get_forecast_file_path(models, file_path, forecast_dates, latest = FALSE)
}
\arguments{
\item{models}{character vector of model abbreviations.}

\item{file_path}{path to local clone of the reichlab/covid19-forecast-hub/data-processed}

\item{forecast_dates}{date vector to generate file paths}

\item{latest}{boolean to generate path to the latest forecast file from each model}
}
\value{
a list of paths to forecast files submitted on a range of forecast dates from selected models
}
\description{
Generate forecast file paths with given data-processed folder path for selected models
}
