% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hub_locations.R
\name{get_locations}
\alias{get_locations}
\title{Create data/hub_locations.rda for loading data from zoltar and covidcast}
\usage{
get_locations()
}
\value{
data frame with columns fips, location_name, population, geo_type,
geo_value, abbreviation
}
\description{
Create data/hub_locations.rda for loading data from zoltar and covidcast
}
