% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_local_zoltar.R
\name{get_model_forecast_history_local_zoltar}
\alias{get_model_forecast_history_local_zoltar}
\title{Get forecast submission history of a specified model from local zoltar sqlite file}
\usage{
get_model_forecast_history_local_zoltar(zoltar_sqlite_file, forecast_model_id)
}
\arguments{
\item{zoltar_sqlite_file}{path to local sqlite file,
either a relative path w.r.t. \code{local_zoltpy_path} or an absolute path.}

\item{forecast_model_id}{one valid id number of a forecast model available in
local zoltar sqlite file. This should come from the result of \code{get_models()}.}
}
\value{
data.frame with \code{id} and \code{forecast_date} columns
}
\description{
Get forecast submission history of a specified model from local zoltar sqlite file
}
