% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_metadata.R
\name{get_model_metadata}
\alias{get_model_metadata}
\title{Get Hub model metadata}
\usage{
get_model_metadata(models = NULL, source = "local_hub_repo", hub_repo_path)
}
\arguments{
\item{models}{optional character vector of model abbreviations for
which to retrieve designations. If not provided, the function returns model
designations for all models.}

\item{source}{string specifying where forecasts will be loaded from:
currently only supports \code{"local_hub_repo"}}

\item{hub_repo_path}{path to local clone of the \code{reichlab/covid19-forecast-hub}
repository}
}
\value{
data.frame with columns corresponding to any fieldnames in the
metadata files from the queried models. NA represents an absence of data
in that corresponding field. Two field names are changed from the raw
metadata field names to align with \code{get_model_designations()} implementation:
\code{model_abbr} is changed to \code{model}
\code{team_model_designation} is changed to \code{designation}
}
\description{
Get Hub model metadata
}
