% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_local_zoltar.R
\name{get_models_local_zoltar}
\alias{get_models_local_zoltar}
\title{Get id and abbrevations of all models available in a specified project
from local zoltar sqlite file}
\usage{
get_models_local_zoltar(zoltar_sqlite_file, project_id)
}
\arguments{
\item{zoltar_sqlite_file}{path to local sqlite file,
either a relative path w.r.t. \code{local_zoltpy_path} or an absolute path.}

\item{project_id}{one valid id number of a project available in
local zoltar sqlite file. This should come from the result of \code{get_zoltar_project_id()}.}
}
\value{
data.frame with \code{id} and \code{model_abbr} columns
}
\description{
Get id and abbrevations of all models available in a specified project
from local zoltar sqlite file
}
