% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_forecast_data.R
\name{get_plot_forecast_data}
\alias{get_plot_forecast_data}
\title{Combine forecasts and truth to plot using plot_forecast().
It will load truth with the corrsponding target and locations in forecast data.}
\usage{
get_plot_forecast_data(
  data,
  horizon,
  quantiles_to_plot,
  location,
  truth_source,
  target_variable,
  truth_as_of = Sys.Date()
)
}
\arguments{
\item{data}{forecasts data frame from load_forecast(). It has columns model,
forecast_date,location, target, type, quantile, value, horizon and
target_end_date.}

\item{horizon}{forecasts are plotted for the horizon time steps after the
forecast date}

\item{quantiles_to_plot}{vector of quanitles to include in the plot}

\item{target_variable}{string specifying target type. It should be one of
"Cumulative Cases","Cumulative Deaths","Incident Cases" and "Incident Deaths."}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the truth_as_of date. Default to today's date}
}
\value{
data frame with columns model,
forecast_date, location, target, type, quantile, value, horizon and
target_end_date.
}
\description{
Combine forecasts and truth to plot using plot_forecast().
It will load truth with the corrsponding target and locations in forecast data.
}
