% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{get_truth_path}
\alias{get_truth_path}
\title{Construct the file path to a truth file}
\usage{
get_truth_path(
  source,
  repo_path,
  target_variable,
  data_location,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{source}{character vector specifying where the truths will
be loaded from: currently support "JHU", "USAFacts", "NYTimes", "HealthData" and "ECDC".}

\item{repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository. Only used when data_location is "local_hub_repo"}

\item{target_variable}{string specifying target type It should be one or more of
"cum death", "inc case", "inc death", "inc hosp".}

\item{data_location}{character specifying the location of truth data.
Currently only supports "local_hub_repo" and "remote_hub_repo".}

\item{hub}{character, which hub to use. Default is "US", other option is
"ECDC"}
}
\value{
character of file path
}
\description{
Construct the file path to a truth file
}
