% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_target_list.R
\name{get_valid_target_list}
\alias{get_valid_target_list}
\title{Create data/hub_targets_us.rda and data/hub_targets_ecdc.rda}
\usage{
get_valid_target_list()
}
\description{
Create data/hub_targets_us.rda and data/hub_targets_ecdc.rda
}
