% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_local_zoltar.R
\name{get_zoltar_project_id_local_zoltar}
\alias{get_zoltar_project_id_local_zoltar}
\title{Get Zoltar project id from local zoltar sqlite file}
\usage{
get_zoltar_project_id_local_zoltar(zoltar_sqlite_file, hub = c("US", "ECDC"))
}
\arguments{
\item{zoltar_sqlite_file}{path to local sqlite file,
either a relative path w.r.t. \code{local_zoltpy_path} or an absolute path.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are \code{"US"} and \code{"ECDC"}}
}
\value{
project id
}
\description{
Get Zoltar project id from local zoltar sqlite file
}
