% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_zoltar_project_url}
\alias{get_zoltar_project_url}
\title{Get Zoltar project URL}
\usage{
get_zoltar_project_url(hub = c("US", "ECDC"), zoltar_connection = NULL)
}
\arguments{
\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}

\item{zoltar_connection}{an authenticated Zoltar connecton}
}
\value{
project url
}
\description{
Get Zoltar project URL
}
