% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_locations_ecdc}
\alias{hub_locations_ecdc}
\title{Available locations in the ECDC Hub}
\format{
A data frame with 32 rows and 3 columns:
\describe{
\item{location_name}{Name of the location}
\item{location}{Location abbreviation}
\item{population}{Location population}
}

A data frame with 54 rows and 4 columns:
\describe{
\item{fips}{FIPS code}
\item{location_name}{Name of the location}
\item{abbreviation}{Location abbreviation}
\item{population}{Location population}
}
}
\usage{
hub_locations_ecdc

hub_locations_ecdc
}
\description{
Data set with available locations for the ECDC hub

Data set with available locations for the FluSight hub
}
\keyword{datasets}
