% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_locations_flusight}
\alias{hub_locations_flusight}
\title{Available locations in the FluSight Hub}
\format{
A data frame with 54 rows and 4 columns:
\describe{
\item{fips}{FIPS code}
\item{location_name}{Name of the location}
\item{abbreviation}{Location abbreviation}
\item{population}{Location population}
}
}
\usage{
hub_locations_flusight
}
\description{
Data set with available locations for the FluSight hub
}
\keyword{datasets}
