% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_targets_ecdc}
\alias{hub_targets_ecdc}
\title{Available targets in the ECDC Hub}
\format{
An object of class \code{character} of length 441.
}
\usage{
hub_targets_ecdc
}
\description{
List with 441 elements
}
\keyword{datasets}
