% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hub_targets_flusight}
\alias{hub_targets_flusight}
\title{Available targets in the FluSight Hub}
\format{
An object of class \code{character} of length 4.
}
\usage{
hub_targets_flusight
}
\description{
List with 4 elements
}
\keyword{datasets}
