% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_with_hub_locations}
\alias{join_with_hub_locations}
\title{Add stored information about location name and population by joining the
data with the correct stored location data}
\usage{
join_with_hub_locations(data, hub = c("US", "ECDC"))
}
\arguments{
\item{data}{data frame to append location data}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}
}
\value{
original data with corresponding location information
}
\description{
Add stored information about location name and population by joining the
data with the correct stored location data
}
