% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_repo.R,
%   R/load_latest_forecasts_repo.R
\name{load_forecast_files_repo}
\alias{load_forecast_files_repo}
\title{Read in a set of forecast files from a repository
without rows having NULLs in value column.}
\usage{
load_forecast_files_repo(
  file_paths,
  locations = NULL,
  types = NULL,
  targets = NULL,
  hub = c("US", "ECDC")
)

load_forecast_files_repo(
  file_paths,
  locations = NULL,
  types = NULL,
  targets = NULL,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{file_paths}{paths to csv forecast files to read in.  It is expected that
the file names are in the format "*YYYY-MM-DD-<model_name>.csv".}

\item{locations}{list of fips. Default to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: \code{"quantile"}
and/or \code{"point"}. Default to all valid forecast types.}

\item{targets}{character vector of targets to retrieve, for example
\code{c('1 wk ahead cum death', '2 wk ahead cum death')}.
Default to \code{NULL} which stands for all valid targets.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are \code{"US"} and \code{"ECDC"}.}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation

data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Read in a set of forecast files from a repository
without rows having NULLs in value column.

Read in a set of forecast files from a repository
without rows having NULLs in value column.
}
\details{
The model_name in the file name will be used for the value of the
model column in the return result.

The model_name in the file name will be used for the value of the
model column in the return result.
}
