% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{load covid forecasts from local files or Zoltar}
\usage{
load_forecasts(
  models,
  last_timezero,
  timezero_window_size,
  types,
  targets,
  source = "hub_repo",
  hub_repo_path
)
}
\arguments{
\item{models}{Character vector of model abbreviations}

\item{last_timezero}{The last timezero date of forecasts to retrieve}

\item{timezero_window_size}{The number of days back to go.  A window size of
0 will retrieve only forecasts submitted on the \code{last_timezero} date.}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death')}

\item{source}{string specifying where forecasts will be loaded from:
currently only "hub_repo" is supported; eventually, either "hub_repo" or
"zoltar"}

\item{hub_repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository}
}
\value{
data frame with columns model, timezero, location, target, quantile,
value
}
\description{
load covid forecasts from local files or Zoltar
}
\details{
One of \code{hub_repo_path} or \code{zoltar_connection} must be provided. If
both are provided, the \code{zoltar_connection} is ignored and forecasts are read
in from the local clone of the repository.
}
