% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_time_period.R
\name{load_forecasts_time_period}
\alias{load_forecasts_time_period}
\title{Load covid forecasts from local hub repo or Zoltar.
Return the most recent forecast from each model within
a specified time period.}
\usage{
load_forecasts_time_period(
  models = NULL,
  start_forecast_date_range,
  end_forecast_date_range,
  locations = NULL,
  types = NULL,
  targets = NULL,
  source = "local_hub_repo",
  hub_repo_path,
  as_of = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)
}
\arguments{
\item{models}{Character vector of model abbreviations.
If missing, forecasts for all models that submitted forecasts
meeting the other criteria are returned.}

\item{start_forecast_date_range}{Start date for specified time period
in 'yyyy-mm-dd' format.}

\item{end_forecast_date_range}{End date for specified time period
in 'yyyy-mm-dd' format.}

\item{locations}{list of fips. Defaults to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to all types  with available forecasts}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets.}

\item{source}{string specifying where forecasts will be loaded from: either
"local_hub_repo" or "zoltar"}

\item{hub_repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository}

\item{as_of}{character for date time to load forecasts submitted as of this time from Zoltar.
Ignored if \code{source} is \code{"local_hub_repo"}.
It could use the format of one of the three examples:
"2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC".
If you would like to set a timezone, it has to be UTC now.
If not, helper function will append the default timezone to your input based on hub parameter.
Default to NULL to load the latest version.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}

\item{verbose}{a boolean for printing messages on zoltar job status. Default to TRUE.}

\item{last_forecast_date}{The forecast date of forecasts to retrieve
in 'yyyy-mm-dd' format.}

\item{data_processed_subpath}{folder within the hub_repo_path that contains
forecast submission files.  Defaults to "data-processed/", which is
appropriate for the covid19-forecast-hub repository.}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Load covid forecasts from local hub repo or Zoltar.
Return the most recent forecast from each model within
a specified time period.
}
