% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_from_hub_repo}
\alias{load_from_hub_repo}
\title{load truth data from a local clone of forecast hub repo}
\usage{
load_from_hub_repo(
  target_variable,
  truth_source,
  repo_path,
  temporal_resolution,
  truth_end_date,
  data_location,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{target_variable}{string specifying target type It should be one or more of
"cum death", "inc case", "inc death", "inc hosp".}

\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support "JHU", "USAFacts", "NYTimes", "HealthData" and "ECDC".}

\item{repo_path}{path to local clone or remote of the corresponding prediction hub repository.}

\item{temporal_resolution}{string specifying temporal resolution
to include: one of 'daily' or 'weekly'}

\item{truth_end_date}{date to include the last available truth point in 'yyyy-mm-dd' format.}

\item{data_location}{character specifying the location of truth data.
Currently only supports "local_hub_repo" or "remote_hub_repo".}

\item{hub}{character, which hub to use. Default is "US", other option is
"ECDC"}
}
\value{
a data frame with columns location, target_end_date, target_variable and value
}
\description{
load truth data from a local clone of forecast hub repo
}
