% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_latest_forecasts_repo.R
\name{load_latest_forecasts_repo}
\alias{load_latest_forecasts_repo}
\title{Load the most recent forecast submitted in a time window
from reichlab/covid19-forecast-hub repo.}
\usage{
load_latest_forecasts_repo(
  file_path,
  models = NULL,
  forecast_dates,
  locations = NULL,
  types = NULL,
  targets = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)
}
\arguments{
\item{file_path}{path to local copy of the data-processed folder of a
forecast hub repo.}

\item{models}{character vector of model abbreviations.
If missing, forecasts for all models that submitted forecasts
meeting the other criteria are returned.}

\item{forecast_dates}{date vector to load the most recent forecast from}

\item{locations}{list of valid location codes. Defaults to all locations with
available forecasts.}

\item{types}{character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to all types  with available forecasts}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets
with available forecasts.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}

\item{verbose}{whether or not to print out diagnostic messages. Default is TRUE}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Load the most recent forecast submitted in a time window
from reichlab/covid19-forecast-hub repo.
}
