% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_latest_forecasts_zoltar.R
\name{load_latest_forecasts_zoltar}
\alias{load_latest_forecasts_zoltar}
\title{Load the most recent forecasts submitted in a time window from zoltar.}
\usage{
load_latest_forecasts_zoltar(models, forecast_dates, locations, types, targets)
}
\arguments{
\item{models}{Character vector of model abbreviations.
If missing, forecasts for all models that submitted forecasts
meeting the other criteria are returned.}

\item{forecast_dates}{date vector to load the most recent forecast from}

\item{locations}{list of valid fips code. Defaults to all locations with
available forecasts.}

\item{types}{character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to c(“quantile”, “point”)}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets.}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value
}
\description{
Load the most recent forecasts submitted in a time window from zoltar.
}
