% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth}
\alias{load_truth}
\title{Load truth data for specified target variable and locations
from one or more truth sources using files in reichlab/covid19-forecast-hub.}
\usage{
load_truth(
  truth_source,
  target_variable,
  truth_end_date,
  temporal_resolution = "weekly",
  locations,
  remote_repo = TRUE,
  local_repo_path = NULL
)
}
\arguments{
\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support "jhu","usafacts", "nytimes"}

\item{target_variable}{string specifying target type It should be one of
"Cumulative Cases","Cumulative Deaths","Incident Cases" and "Incident Deaths"}

\item{truth_end_date}{date to include the last available truth point}

\item{temporal_resolution}{character specifying temporal resolution
to include: currently support "weekly" and "daily". Default to 'weekly'}

\item{locations}{vector of valid fips code. Defaults to all locations with available forecasts.}

\item{remote_repo}{boolean specifying whether repo is remote or local}

\item{local_repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository. Only used when remote_repo is FALSE}
}
\value{
data frame with columns model, target, target_end_date, location and value
}
\description{
Load truth data for specified target variable and locations
from one or more truth sources using files in reichlab/covid19-forecast-hub.
}
