% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth_zoltar}
\alias{load_truth_zoltar}
\title{load truth data from zoltar server}
\usage{
load_truth_zoltar(
  target_variable,
  locations = NULL,
  as_of = NULL,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{target_variable}{string specifying target type It should be one or more of
\code{"cum death"} and \code{"inc death"}.}

\item{locations}{vector of valid location code.
Default to \code{NULL}, all locations with available forecasts.
US hub is using FIPS code and ECDC hub is using country name abbreviation.}

\item{as_of}{optional character vector of "as of" dates to use for querying truths in
format 'yyyy-mm-dd'. For each spatial unit and temporal reporting unit, the last
available data with an issue date on or before the given \code{as_of} date are returned.
Default to \code{NULL}, the latest available date.}

\item{hub}{character, which hub to use. Default is \code{"US"}.
\code{"ECDC"} is not available now.}
}
\value{
data.frame with columns \code{model}, \code{location}, \code{target_end_date}, \code{target_variable}
and \code{value}.
}
\description{
load truth data from zoltar server
}
