% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_to_fips}
\alias{name_to_fips}
\title{Get FIPS or CBSA codes from county or metropolitan area locations}
\usage{
name_to_fips(data, hub = c("US", "ECDC"))
}
\arguments{
\item{data}{a vector of strings for fips code, CBSA codes, location names
such as "Hampshire County, MA", "Alabama", "United Kingdom".
A US county location names must include state abbreviation.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are \code{"US"}, \code{"ECDC"} and \code{"FluSight"}}
}
\value{
A vector of FIPS or CBSA codes
}
\description{
Get FIPS or CBSA codes from county or metropolitan area locations
}
