% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecasts with truth for only one model, one target and one location}
\usage{
plot_forecast(
  forecast_data,
  model,
  target_variable,
  location,
  intervals = c(0.5, 0.8, 0.95),
  horizon,
  truth_source = "jhu",
  truth_as_of
)
}
\arguments{
\item{forecast_data}{data frame with truth and forecasts from load_forecasts()}

\item{model}{model_abbr specifying model to plot}

\item{target_variable}{string specifying target type. It should be one of
"Cumulative Cases","Cumulative Deaths","Incident Cases" and "Incident Deaths,"}

\item{location}{string for fips code or 'US'}

\item{intervals}{values indicating which central prediction interval levels
to plot, defaults to c(.5, .8, .95)}

\item{horizon}{forecasts are plotted for the horizon time steps after the
forecast date}

\item{truth_source}{character specifying where the truths will be loaded from.
Currently supports "jhu","usafacts", "nytimes". Default to "jhu".}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the truth_as_of date}
}
\value{
ggplot graph
}
\description{
Plot forecasts with truth for only one model, one target and one location
}
