% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_hospitalization}
\alias{preprocess_hospitalization}
\title{Preprocess raw hospitalization data into Cumulative/Incident hospitalizations and write to CSVs}
\usage{
preprocess_hospitalization(save_location = "./data-truth")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./data-truth"}
}
\value{
data frame of cumulative and incident hospitalization data
}
\description{
Preprocess raw hospitalization data into Cumulative/Incident hospitalizations and write to CSVs
}
