% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_jhu}
\alias{preprocess_jhu}
\title{Preprocess raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases and write to CSVs}
\usage{
preprocess_jhu(save_location = "./data-truth")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./data-truth"}
}
\description{
Preprocess raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases and write to CSVs
}
