% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zoltar_truth.R
\name{preprocess_truth_for_zoltar}
\alias{preprocess_truth_for_zoltar}
\title{Generate versioned truth data for zoltar for a specified target
It only includes national and state-level truth data.}
\usage{
preprocess_truth_for_zoltar(target, issue_date = NULL)
}
\arguments{
\item{target}{string specifying target.
Currently only support "Cumulative Deaths" and "Incident Deaths".}

\item{issue_date}{optional date specifying issue date of truth data.
Default to NULL which will load the latest truth data from covidData.}
}
\value{
data frame with timezero, unit, target and value
}
\description{
Generate versioned truth data for zoltar for a specified target
It only includes national and state-level truth data.
}
