% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_usafacts}
\alias{preprocess_usafacts}
\title{Preprocess raw truth data from USAFacts into Cumulative/Incident - Deaths/Cases and write to CSVs}
\usage{
preprocess_usafacts(save_location = "./data-truth/usafacts/")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to \code{"./data-truth/usafacts/raw/"}}
}
\description{
Preprocess raw truth data from USAFacts into Cumulative/Incident - Deaths/Cases and write to CSVs
}
