% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_visualization_truth}
\alias{preprocess_visualization_truth}
\title{Preprocess raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases for visualization
purpose and write to JSON files}
\usage{
preprocess_visualization_truth(
  save_location = "./visualization/vis-master/covid-csv-tools/dist/truth"
)
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./visualization/vis-master/covid-csv-tools/dist/truth"}
}
\description{
Preprocess raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases for visualization
purpose and write to JSON files
}
