% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_forecasts.R
\name{score_forecasts}
\alias{score_forecasts}
\title{Score forecasts}
\usage{
score_forecasts(forecasts, truth, desired_score_types, return_format = "wide")
}
\arguments{
\item{forecasts}{required data.frame with forecasts in the format returned
by load_forecasts}

\item{desired_score_types}{character vector of scores to calculate; defaults to returning
all available scores}

\item{return_format}{string: "long" returns long format with a column for
"score_name" and a column for "score_value"; "wide" returns wide format with
a separate column for each score. Defaults to "wide".}

\item{truths}{required data.frame with forecasts in the format returned
by load_truth}
}
\value{
data.frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, horizon, target_end_date, and scores:
If return_format is "long", also contains columns score_name and score_value
where score_name is the type of score calculated and score_value has the numeric
value of the score.
If return_format is "wide", each calculated score is in its own column.
}
\description{
Score forecasts
}
