% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_zoltar_connection}
\alias{setup_zoltar_connection}
\title{Set up Zoltar connection}
\usage{
setup_zoltar_connection(staging = FALSE)
}
\arguments{
\item{staging}{boolean to change Zoltar server for staging}
}
\value{
a Zoltar connection
}
\description{
Set up Zoltar connection
}
