% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zoltar_truth.R
\name{zoltar_truth_data}
\alias{zoltar_truth_data}
\title{Generate the most up to date truth data for zoltar
It only includes national and state-level truth data.}
\usage{
zoltar_truth_data(save_location = "./data-truth")
}
\arguments{
\item{save_location}{character specifying the location of to save zoltar truth data.
Default to "./data-truth"}
}
\description{
Generate the most up to date truth data for zoltar
It only includes national and state-level truth data.
}
