% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_weekly.R
\name{aggregate_to_weekly}
\alias{aggregate_to_weekly}
\title{Aggregate daily data to weekly data}
\usage{
aggregate_to_weekly(data)
}
\arguments{
\item{data}{data frame with columns model, location, target_variable, target_end_date and value}
}
\value{
data frame with the same set of input columns and weekly aggregated data in value column
}
\description{
If the last week is not complete, drop all observations from the
previous Saturday in that week
}
