% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidHubUtils-package.R
\docType{package}
\name{covidHubUtils}
\alias{covidHubUtils}
\alias{covidHubUtils-package}
\title{covidHubUtils - Utility functions for the COVID-19 forecast hub}
\description{
Utility functions for the COVID-19 forecast hub
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/reichlab/covidHubUtils}
}

}
\author{
Evan L. Ray \href{mailto:elray@umass.edu}{elray@umass.edu}
}
