% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_management.R
\name{date_to_week_end_date}
\alias{date_to_week_end_date}
\title{Calculate week end date (i.e., Saturday) for the MMWR epidemic week that is
offset a specified number of epidemic week from a specified date}
\usage{
date_to_week_end_date(timezero, week_offset = 0)
}
\arguments{
\item{timezero}{character vector of dates in 'yyyy-mm-dd' format}

\item{integer}{vector of week offsets.  must be either length 1 or the same
length as timezero}
}
\value{
character vector of dates in 'yyyy-mm-dd' format
}
\description{
Calculate week end date (i.e., Saturday) for the MMWR epidemic week that is
offset a specified number of epidemic week from a specified date
}
