% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{download_raw_nytimes}
\alias{download_raw_nytimes}
\title{Download historical raw truth data from NYTimes and write to CSV files. Note:
updates to the NYTimes data stopped in March 2023.}
\usage{
download_raw_nytimes(save_location = "./data-truth/nytimes/raw/")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to \code{"./data-truth/nytimes/raw/"}}
}
\value{
data.frame of national, state and county level raw truth data
}
\description{
Download historical raw truth data from NYTimes and write to CSV files. Note:
updates to the NYTimes data stopped in March 2023.
}
