% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_forecast_data.R
\name{get_plot_forecast_data}
\alias{get_plot_forecast_data}
\title{Combine \code{\link[=load_truth]{load_truth()}} and \code{\link[=pivot_forecasts_wider]{pivot_forecasts_wider()}}}
\usage{
get_plot_forecast_data(
  forecast_data,
  truth_data = NULL,
  models_to_plot,
  forecast_dates_to_plot,
  horizons_to_plot,
  quantiles_to_plot,
  locations_to_plot = NULL,
  plot_truth = TRUE,
  truth_source,
  target_variable_to_plot,
  truth_as_of = NULL,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{forecast_data}{required data.frame with forecasts in the format returned
by \code{\link[=load_forecasts]{load_forecasts()}}.
It has columns \code{model}, \code{forecast_date}, \code{location}, \code{target}, \code{type}, \code{quantile},
\code{value}, \code{horizon} and \code{target_end_date}.}

\item{truth_data}{optional data.frame from one truth source in the format returned
by \code{\link[=load_truth]{load_truth()}}. It needs to have columns \code{model}, \code{target_variable},
\code{target_end_date}, \code{location} and \code{value}.
\code{Model} column can be "Observed Data (a truth source)".}

\item{models_to_plot}{characters of model abbreviations}

\item{forecast_dates_to_plot}{date string vectors for forecast dates to plot.
Default to all forecast dates available in \code{forecast_data}.}

\item{horizons_to_plot}{forecasts are plotted for the horizon time steps after
the forecast date.}

\item{quantiles_to_plot}{vector of quantiles to include in the plot}

\item{locations_to_plot}{a vector of strings for fips code, CBSA codes, location names
such as "Hampshire COunty, MA", "United Kingdom","Alabama"
for a US county location names must include state abbreviation.
Default to all locations available in \code{forecast_data}.}

\item{plot_truth}{logical to indicate whether truth data should be plotted.
Default to \code{TRUE}.}

\item{truth_source}{character specifying where the truth data will
be loaded from if truth_data is not provided. Currently support \code{"JHU"},
\code{"USAFacts"}, \code{"NYTimes"} and \code{"HealthData"}.
Optional if \code{truth_data} is provided.}

\item{target_variable_to_plot}{string specifying target type. It should be one of
\code{"cum death"}, \code{"inc case"}, \code{"inc death"} and \code{"inc hosp"}.}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the \code{truth_as_of} date.}

\item{hub}{character, which hub to use. Default is \code{"US"}, other option is
\code{"ECDC"}}
}
\value{
data.frame with columns \code{model},
\code{forecast_date}, \code{location}, \code{target_variable}, \code{type}, \code{quantile}, \code{value},
\code{horizon} and \code{target_end_date}.
}
\description{
Combine \code{\link[=load_truth]{load_truth()}} and \code{\link[=pivot_forecasts_wider]{pivot_forecasts_wider()}}
}
