% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{get_truth_path}
\alias{get_truth_path}
\title{Construct the file path to a truth file}
\usage{
get_truth_path(
  source,
  repo_path,
  target_variable,
  data_location,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{source}{character vector specifying where the truths will
be loaded from: currently support \code{"JHU"}, \code{"USAFacts"}, \code{"NYTimes"}, \verb{"HealthData}" and \code{"ECDC"}.}

\item{repo_path}{path to local clone or remote of the corresponding prediction hub repository.}

\item{target_variable}{string specifying target type It should be one or more of
\code{"cum death"}, \code{"inc case"}, \code{"inc death"}, \code{"inc hosp"}.}

\item{data_location}{character specifying the location of truth data.
Currently only supports \code{"local_hub_repo"} and \code{"remote_hub_repo"}.}

\item{hub}{character, which hub to use. Default is \code{"US"}, other option is
\code{"ECDC"}}
}
\value{
character of file path
}
\description{
Construct the file path to a truth file
}
