% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_latest_forecasts_repo.R
\name{load_forecast_files_repo}
\alias{load_forecast_files_repo}
\title{Read in a set of forecast files from a repository}
\usage{
load_forecast_files_repo(
  file_paths,
  locations = NULL,
  types = NULL,
  targets = NULL
)
}
\arguments{
\item{file_paths}{paths to csv forecast files to read in.  It is expected that
the file names are in the format "*YYYY-MM-DD-<model_name>.csv".}

\item{locations}{list of valid fips codes. Defaults to all locations with
available forecasts.}

\item{types}{character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to all types in the forecast file}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets
in the forecast file.}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Read in a set of forecast files from a repository
}
\details{
The model_name in the file name will be used for the value of the
model column in the return result.
}
