% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Load all available forecasts submitted on forecast_dates from Zoltar
or local hub repo.}
\usage{
load_forecasts(
  models = NULL,
  forecast_dates = NULL,
  locations = NULL,
  types = NULL,
  targets = NULL,
  source = "zoltar",
  hub_repo_path,
  as_of = NULL,
  verbose = TRUE,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{models}{Character vector of model abbreviations.
Default all models that submitted forecasts meeting the other criteria.}

\item{forecast_dates}{The forecast date of forecasts to retrieve.
Default to all valid forecast dates.
The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.}

\item{locations}{list of location codes Default to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: “quantile”
or “point”. Default to all valid forecast types.}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death').
Default to NULL which stands for all valid targets.}

\item{source}{string specifying where forecasts will be loaded from: either
"local_hub_repo" or "zoltar". Default to "zoltar"}

\item{hub_repo_path}{path to local clone of the forecast hub
repository}

\item{as_of}{a date in YYYY-MM-DD format to load forecasts submitted as of this date.
Default to NULL to load the latest version.}

\item{verbose}{a boolean for printing messages on zoltar job status poll}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Load all available forecasts submitted on forecast_dates from Zoltar
or local hub repo.
}
