% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Load all available forecasts submitted around forecast dates from Zoltar
or local hub repo.}
\usage{
load_forecasts(
  models = NULL,
  dates = NULL,
  date_window_size = 0,
  locations = NULL,
  types = NULL,
  targets = NULL,
  source = "zoltar",
  hub_repo_path,
  data_processed_subpath = "data-processed/",
  as_of = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)
}
\arguments{
\item{models}{Character vector of model abbreviations.
Default all models that submitted forecasts meeting the other criteria.}

\item{dates}{The forecast date of forecasts to retrieve.
A vector of one or more Date objects or character strings in format “YYYY-MM-DD”
Default to all valid forecast dates.}

\item{date_window_size}{The number of days across which to
look for the most recent forecasts for each date in dates parameter.
Default to 0, which means to only look at the dates parameter only.}

\item{locations}{list of location codes. Default to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: \code{"quantile"}
and/or \code{"point"}. Default to all valid forecast types.}

\item{targets}{character vector of targets to retrieve, for example
\code{c('1 wk ahead cum death', '2 wk ahead cum death')}.
Default to \code{NULL} which stands for all valid targets.}

\item{source}{string specifying where forecasts will be loaded from: either
\code{"local_hub_repo"} or \code{"zoltar"}. Default to \code{"zoltar"}.}

\item{hub_repo_path}{path to local clone of the forecast hub
repository}

\item{data_processed_subpath}{folder within the hub_repo_path that contains
forecast submission files.  Default to \code{"data-processed/"}, which is
appropriate for the covid19-forecast-hub repository.}

\item{as_of}{character for date time to load forecasts submitted as of this time from Zoltar.
Ignored if \code{source = "local_hub_repo"}.
It could use the format of one of the three examples:
\verb{"2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC"}.
If you would like to set a timezone, it has to be UTC now.
If not, helper function will append the default timezone to your input based on \code{hub} parameter.
Default to NULL to load the latest version.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are \code{"US"} and \code{"ECDC"}.}

\item{verbose}{logical to print out diagnostic messages. Default is \code{TRUE}.}
}
\value{
data.frame with columns \code{model}, \code{forecast_date}, \code{location}, \code{horizon},
\code{temporal_resolution}, \code{target_variable}, \code{target_end_date}, \code{type}, \code{quantile}, \code{value},
\code{location_name}, \code{population}, \code{geo_type}, \code{geo_value}, \code{abbreviation}
}
\description{
If \code{date_window_size} is 0, this function returns all available forecasts
submitted on every day in \code{dates}.
}
\details{
If \code{date_window_size} is not 0, this function will look for all the latest
forecasts that are submitted within window size for each day in \code{dates}.
}
\examples{
# load forecasts from US forecast hub
load_forecasts(
  models = "COVIDhub-ensemble",
  dates = "2020-12-07",
  locations = "US",
  types = c("point", "quantile"),
  targets = paste(1:4, "wk ahead inc case"),
  source = "zoltar",
  verbose = FALSE,
  as_of = NULL
)

# load forecasts from ECDC forecast hub
load_forecasts(
  models = "ILM-EKF",
  hub = c("ECDC", "US"),
  dates = "2021-03-08",
  locations = "GB",
  targets = paste(1:4, "wk ahead inc death"),
  source = "zoltar"
)
}
