% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Load covid forecasts from Zoltar.
Return all available forecasts submitted on forecast_dates}
\usage{
load_forecasts(
  models = NULL,
  forecast_dates = NULL,
  locations = NULL,
  types = NULL,
  targets = NULL
)
}
\arguments{
\item{models}{Character vector of model abbreviations.
If missing, forecasts for all models that submitted forecasts
meeting the other criteria are returned.}

\item{forecast_dates}{The forecast date of forecasts to retrieve.
Defaults to all valid forecast dates in Zoltar.
The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.}

\item{locations}{list of fips. Defaults to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to c(“quantile”, “point”)}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets.}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
The function will throw a warning and return an empty data frame when
no forecasts are submitted on any dates in forecast_dates for selected models,
locations, types and target.
}
