% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts.R
\name{load_forecasts}
\alias{load_forecasts}
\title{Load all available forecasts submitted on forecast_dates from Zoltar
or local hub repo.}
\usage{
load_forecasts(
  models = NULL,
  forecast_dates = NULL,
  locations = NULL,
  types = NULL,
  targets = NULL,
  source = "zoltar",
  hub_repo_path,
  data_processed_subpath = "data-processed/",
  as_of = NULL,
  hub = c("US", "ECDC"),
  verbose = TRUE
)
}
\arguments{
\item{models}{Character vector of model abbreviations.
Default all models that submitted forecasts meeting the other criteria.}

\item{forecast_dates}{The forecast date of forecasts to retrieve.
Default to all valid forecast dates.
The function will throw an error if all dates in this parameter are invalid forecast dates in Zoltar.}

\item{locations}{list of location codes. Default to all locations with available forecasts.}

\item{types}{Character vector specifying type of forecasts to load: “quantile”
or “point”. Default to all valid forecast types.}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death').
Default to NULL which stands for all valid targets.}

\item{source}{string specifying where forecasts will be loaded from: either
"local_hub_repo" or "zoltar". Default to "zoltar"}

\item{hub_repo_path}{path to local clone of the forecast hub
repository}

\item{data_processed_subpath}{folder within the hub_repo_path that contains
forecast submission files.  Defaults to "data-processed/", which is
appropriate for the covid19-forecast-hub repository.}

\item{as_of}{character for date time to load forecasts submitted as of this time from Zoltar.
Ignored if \code{source} is \code{"local_hub_repo"}.
It could use the format of one of the three examples:
"2021-01-01", "2020-01-01 01:01:01" and "2020-01-01 01:01:01 UTC".
If you would like to set a timezone, it has to be UTC now.
If not, helper function will append the default timezone to your input based on hub parameter.
Default to NULL to load the latest version.}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are "US" and "ECDC"}

\item{verbose}{whether or not to print out diagnostic messages. Default is TRUE}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value,
location_name, population, geo_type, geo_value, abbreviation
}
\description{
Load all available forecasts submitted on forecast_dates from Zoltar
or local hub repo.
}
\examples{
\dontrun{
forecasts_US <- load_forecasts(models = "COVIDhub-ensemble",
 forecast_date = "2020-12-07",
 locations = "US",
 types = c("point","quantile"),
 targets = paste(1:4, "wk ahead inc case"),
 source = "zoltar",
 verbose = FALSE,
 as_of = NULL)

forecasts_ECDC <- load_forecasts(models = "ILM-EKF",
 hub = c("ECDC","US"),
 forecast_date = "2021-03-08",
 locations = "GB",
 targets = paste(1:4, "wk ahead inc death"),
 source = "zoltar")
}

}
