% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_repo.R
\name{load_forecasts_repo}
\alias{load_forecasts_repo}
\title{Load forecasts from reichlab/covid19-forecast-hub repo.}
\usage{
load_forecasts_repo(
  file_path,
  models,
  forecast_dates,
  locations,
  types,
  targets
)
}
\arguments{
\item{file_path}{path to local clone of the reichlab/covid19-forecast-hub
repository}

\item{models}{character vector of model abbreviations.
If missing, forecasts for all models that submitted forecasts
meeting the other criteria are returned.}

\item{forecast_dates}{date vector to load the most recent forecast from}

\item{locations}{list of valid fips code. Defaults to all locations with
available forecasts.}

\item{types}{character vector specifying type of forecasts to load: “quantile”
or “point”. Defaults to c(“quantile”, “point”)}

\item{targets}{character vector of targets to retrieve, for example
c('1 wk ahead cum death', '2 wk ahead cum death'). Defaults to all targets.}
}
\value{
data frame with columns model, forecast_date,location, target,
type, quantile, value, horizon and target_end_date.
}
\description{
Load forecasts from reichlab/covid19-forecast-hub repo.
}
