% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_from_hub_repo}
\alias{load_from_hub_repo}
\title{load truth data from a local clone of forecast hub repo}
\usage{
load_from_hub_repo(
  target_variable,
  truth_source,
  repo_path,
  temporal_resolution,
  truth_end_date,
  data_location,
  hub = c("US", "ECDC", "FluSight")
)
}
\arguments{
\item{target_variable}{string specifying target type It should be one or more of
\code{"cum death"}, \code{"inc case"}, \code{"inc death"}, \code{"inc hosp"}.}

\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support \code{"JHU"}, \code{"NYTimes"}, \code{"HealthData"} and \code{"ECDC"}.}

\item{repo_path}{path to local clone or remote of the corresponding prediction hub repository.}

\item{temporal_resolution}{string specifying temporal resolution
to include: one of \code{'daily'} or \code{'weekly'}}

\item{truth_end_date}{date to include the last available truth point in \code{'yyyy-mm-dd'} format.}

\item{data_location}{character specifying the location of truth data.
Currently only supports \code{"local_hub_repo"} or \code{"remote_hub_repo"}.}

\item{hub}{character, which hub to use. Default is "US".
Other options are "ECDC" and "FluSight".}
}
\value{
a data.frame with columns \code{location}, \code{target_end_date}, \code{target_variable} and \code{value}
}
\description{
load truth data from a local clone of forecast hub repo
}
