% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth}
\alias{load_truth}
\title{Load truth data for specified target variable and locations
from one or more truth sources using files in reichlab/covid19-forecast-hub.}
\usage{
load_truth(
  truth_source,
  target_variable,
  truth_end_date = Sys.Date(),
  temporal_resolution = "weekly",
  locations,
  data_location = "remote_hub_repo",
  local_repo_path = NULL
)
}
\arguments{
\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support "JHU","USAFacts", "NYTimes"}

\item{target_variable}{string specifying target type It should be one of
"Cumulative Cases","Cumulative Deaths","Incident Cases" and "Incident Deaths"}

\item{truth_end_date}{date to include the last available truth point. Default to system date.}

\item{temporal_resolution}{character specifying temporal resolution
to include: currently support "weekly" and "daily". Default to 'weekly'.}

\item{locations}{vector of valid fips code. Defaults to all locations with available forecasts.}

\item{data_location}{character specifying the location of truth data.
Currently only supports "local_hub_repo" and "remote_hub_repo". Default to "remote_hub_repo".}

\item{local_repo_path}{path to local clone of the reichlab/covid19-forecast-hub
repository. Only used when data_location is "local_hub_repo"}
}
\value{
data frame with columns model, inc_cum, death_case, target_end_date, location and value
}
\description{
Load truth data for specified target variable and locations
from one or more truth sources using files in reichlab/covid19-forecast-hub.
}
