% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth}
\alias{load_truth}
\title{Load truth data under multiple target variables
from multiple truth sources
using files in \code{reichlab/covid19-forecast-hub}.}
\usage{
load_truth(
  truth_source = NULL,
  target_variable = NULL,
  as_of = NULL,
  truth_end_date = NULL,
  temporal_resolution = NULL,
  locations = NULL,
  data_location = NULL,
  local_repo_path = NULL,
  hub = c("US", "ECDC")
)
}
\arguments{
\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support \code{"JHU"}, \code{"USAFacts"}, \code{"NYTimes"}, \verb{"HealthData}" and \code{"ECDC"}.
If \code{NULL}, default for US hub is \code{c("JHU", "HealthData")}.
If \code{NULL}, default for ECDC hub is \code{c("JHU")}.}

\item{target_variable}{string specifying target type It should be one or more of
\code{"cum death"}, \code{"inc case"}, \code{"inc death"}, \code{"inc hosp"}.
If \code{NULL}, default for US hub is \code{c("inc case", "inc death", "inc hosp")}.
If \code{NULL}, default for ECDC hub is \code{c("inc case", "inc death")}.}

\item{as_of}{character vector of "as of" dates to use for querying truths in
format 'yyyy-mm-dd'. For each spatial unit and temporal reporting unit, the last
available data with an issue date on or before the given \code{as_of} date are returned.
This is only available for \code{covidData} and \code{zoltar}now.}

\item{truth_end_date}{date to include the last available truth point in 'yyyy-mm-dd' format.
If \code{NULL},default to system date.}

\item{temporal_resolution}{character specifying temporal resolution
to include: currently support \code{"weekly"} and \code{"daily"}.
If \code{NULL}, default to \code{"weekly"} for cases and deaths, \code{"daily"} for hospitalizations.
Weekly \code{temporal_resolution} will not be applied to \code{"inc hosp"} when
multiple target variables are specified.
\code{"ECDC"} truth data is weekly by default. Daily level data is not available.}

\item{locations}{vector of valid location code.
If \code{NULL}, default to all locations with available forecasts.
US hub is using FIPS code and ECDC hub is using country name abbreviation.}

\item{data_location}{character specifying the location of truth data.
Currently only supports \code{"local_hub_repo"}, \code{"remote_hub_repo"}, \code{"covidData"} and \code{"zoltar"}.
If \code{NULL}, default to \code{"remote_hub_repo"}.}

\item{local_repo_path}{path to local clone of the \code{reichlab/covid19-forecast-hub}
repository. Only used when data_location is \code{"local_hub_repo"}}

\item{hub}{character, which hub to use. Default is \code{"US"}, other option is
\code{"ECDC"}. \code{"ECDC"} is not available when \code{data_location = "zoltar"}.}
}
\value{
data.frame with columns \code{model}, \code{target_variable}, \code{target_end_date},
\code{location}, \code{value}, \code{location_name}, \code{population}, \code{geo_type}, \code{geo_value}, \code{abbreviation}
and \code{full_location_name}
}
\description{
\code{"inc hosp"} is only available from \code{"HeatlthData"} and this function is not loading
data for other target variables from \code{"HealthData"}.
}
\details{
When loading data for multiple \code{target_variables}, \code{temporal_resolution} will be applied
to all target variables but \code{"inc hosp"}. In that case, this function will return
daily incident hospitalization counts along with other data.

Weekly temporal resolution will be applied to \code{"inc hosp"} if the user specifies \code{"inc hosp"}
as the only \code{target_variable}.

When loading weekly data, if there are not enough observations for a week, the corresponding
weekly count would be \code{NA} in resulting data frame.
}
\examples{
library(covidHubUtils)

# load for US
load_truth(
  truth_source = c("JHU", "HealthData"),
  target_variable = c("inc case", "inc death", "inc hosp")
)

# load for ECDC
load_truth(
  truth_source = c("JHU"),
  target_variable = c("inc case", "inc death"),
  hub = "ECDC"
)
}
