% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_truth.R
\name{load_truth}
\alias{load_truth}
\title{Load truth data under multiple target variables from multiple truth sources}
\usage{
load_truth(
  truth_source = NULL,
  target_variable = NULL,
  as_of = NULL,
  truth_end_date = NULL,
  temporal_resolution = NULL,
  locations = NULL,
  data_location = NULL,
  local_repo_path = NULL,
  hub = c("US", "ECDC", "FluSight")
)
}
\arguments{
\item{truth_source}{character vector specifying where the truths will
be loaded from: currently support \code{"JHU"}, \code{"USAFacts"}, \code{"NYTimes"}, \verb{"HealthData}" and \code{"ECDC"}.
If \code{NULL}, default for US hub is \code{c("JHU", "HealthData")}.
If \code{NULL}, default for ECDC hub is \code{c("JHU")}.
If \code{NULL}, default for FluSight hub is \code{c("HealthData")}.}

\item{target_variable}{string specifying target type It should be one or more of
\code{"cum death"}, \code{"inc case"}, \code{"inc death"}, \code{"inc hosp"}.
If \code{NULL}, default for US hub is \code{c("inc case", "inc death", "inc hosp")}.
If \code{NULL}, default for ECDC hub is \code{c("inc case", "inc death")}.
If \code{NULL}, default for FluSight hub is \code{c("inc flu hosp")}.}

\item{as_of}{character vector of "as of" dates to use for querying truths in
format 'yyyy-mm-dd'. For each spatial unit and temporal reporting unit, the last
available data with an issue date on or before the given \code{as_of} date are returned.
This is only available for \code{covidData} now.}

\item{truth_end_date}{date to include the last available truth point in 'yyyy-mm-dd' format.
If \code{NULL},default to system date.}

\item{temporal_resolution}{character specifying temporal resolution
to include: currently support \code{"weekly"} and \code{"daily"}.
If \code{NULL}, default to \code{"weekly"} for cases and deaths, \code{"daily"} for hospitalizations.
Weekly \code{temporal_resolution} will not be applied to \code{"inc hosp"} and \code{"inc fluhosp"}when
multiple target variables are specified.
\code{"ECDC"} truth data is weekly by default. Daily level data is not available.}

\item{locations}{a vector of strings for fips code, CBSA codes, location names
such as "Hampshire COunty, MA", "United Kingdom","Alabama"
for a US county location names must include state abbreviation.
If \code{NULL}, default to all locations with available forecasts.}

\item{data_location}{character specifying the location of truth data.
Currently only supports \code{"local_hub_repo"}, \code{"remote_hub_repo"} and \code{"covidData"}.
If \code{NULL}, default to \code{"remote_hub_repo"}.}

\item{local_repo_path}{path to local clone of the hub repository.
Only used when data_location is \code{"local_hub_repo"}}

\item{hub}{character, which hub to use. Default is "US".
Other options are "ECDC" and "FluSight".}
}
\value{
data.frame with columns \code{model}, \code{target_variable}, \code{target_end_date},
\code{location}, \code{value}, \code{location_name}, \code{population} and extra information in these cases
\itemize{
\item If \code{hub = "US"}, it returns extra columns \code{geo_type}, \code{geo_value}, \code{abbreviation} and \code{full_location_name}.
\item If \code{truth_source = "ECDC"}, this function returns extra columns \code{week_start}. However, when \code{target_variable} is only
\verb{inc hosp}, there are no extra columns appended to the resulting data frame.
}
}
\description{
By default, for the US hub, the resulting data.frame contains data for weekly incident cases (JHU),
weekly incident deaths (JHU) and daily incident hospitalization (HealthData) at all county, state and
national level. For the ECDC hub, the default resulting data.frame contains data for weekly incident cases (JHU),
weekly incident deaths (JHU) and daily incident hospitalization (ECDC) for all European countries.
For the FluSight hub, the default resulting data.frame contains data for weekly incident hospitalization (HealthData)
for all US locations.
}
\details{
\itemize{
\item\code{"inc hosp"} is only available from \code{"HealthData"} and \code{"ECDC"}.\code{"inc flu hosp"} is only available from \code{"HealthData"}.

\item This function is not loading data for other target variables from \code{"HealthData"}.

\item When loading data for multiple target variables for the US hub, \code{temporal_resolution} will be applied
to all target variables but \code{"inc hosp"} and \code{"inc flu hosp"}. In that case, the function will return
daily incident COVID hospitalization counts and weekly incident Influenza hospitalization.

\item For the US hub, weekly temporal resolution will be applied to \code{"inc hosp"} if the user specifies \code{"inc hosp"}
as the only \code{target_variable}.On the other hand, \code{temporal_resolution} will
be applied to \code{"inc hosp"} in all cases for the ECDC hub.

\item When aggregating daily data, if there are not enough observations for a week, the corresponding
weekly count would be \code{NA} in resulting data frame.

\item \code{as_of} is only supported when \code{data_location = "covidData"}. Otherwise, this function
will return a warning.
}
}
\examples{
library(covidHubUtils)

# load for US
load_truth(
  truth_source = c("JHU", "HealthData"),
  target_variable = c("inc case", "inc death", "inc hosp")
)

# load for ECDC
load_truth(
  truth_source = c("JHU"),
  target_variable = c("inc case", "inc death"),
  hub = "ECDC"
)
}
