% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_to_fips}
\alias{name_to_fips}
\title{Get FIPS or CBSA codes from county or metropolitan area locations}
\usage{
name_to_fips(data, hub = c("US", "ECDC"))
}
\arguments{
\item{data}{data frame to append location data}

\item{hub}{character vector, where the first element indicates the hub
from which to load forecasts. Possible options are \code{"US"} and \code{"ECDC"}}
}
\value{
A vector of FIPS or CBSA codes
}
\description{
Get FIPS or CBSA codes from county or metropolitan area locations
}
