% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_forecasts_wider.R
\name{pivot_forecasts_wider}
\alias{pivot_forecasts_wider}
\title{Return a data frame of forecasts with predictive quantiles in columns.}
\usage{
pivot_forecasts_wider(forecast_data, quantiles = NULL)
}
\arguments{
\item{forecast_data}{required data.frame with forecasts in the format returned
by \code{\link[=load_forecasts]{load_forecasts()}}.
It has columns \code{model},\code{forecast_date},\code{location}, \code{target}, \code{type},
\code{quantile}, \code{value}, \code{horizon} and \code{target_end_date}.}

\item{quantiles}{vector of quantiles to return for plotting
Default to all available quantiles in \code{forecast_data}.
Note that point forecasts and median quantiles do not necessarily align.
Throws an error if quantile levels are not in the forecast data.}
}
\value{
data.frame with columns \code{model}, \code{forecast_date}, \code{location},
\code{horizon}, \code{temporal_resolution}, \code{target_variable}, \code{target_end_date}, \code{type}, \code{point},
\code{point_type}, \verb{Prediction Interval}, \code{upper}, \code{lower}.
}
\description{
Return a data frame of forecasts with predictive quantiles in columns.
}
