% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_forecasts_wider.R
\name{pivot_forecasts_wider}
\alias{pivot_forecasts_wider}
\title{Return a data frame of forecasts with predictive quantiles in columns.}
\usage{
pivot_forecasts_wider(
  data,
  quantiles = c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975)
)
}
\arguments{
\item{data}{forecasts data frame from load_forecast(). It has columns model,
forecast_date,location, target, type, quantile, value, horizon and
target_end_date.}

\item{quantiles}{vector of quantiles to return for plotting
Defaults to c(0.025, 0.1, 0.25, 0.5, 0.75, 0.9, 0.975).
Note that point forecasts and median quantiles do not necessarily align.
Throws an error if quantile levels are not in the forecast data.}
}
\value{
data frame with columns model, forecast_date, location, target,
horizon, type, point, target_end_date, point_forecast_type, alpha, Prediction Interval,
upper, lower.
}
\description{
Return a data frame of forecasts with predictive quantiles in columns.
}
