% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecasts with truth for only one model, one target and one location}
\usage{
plot_forecast(
  forecast_data,
  truth_data = NULL,
  models,
  target_variable,
  locations,
  facet = NULL,
  facet_scales = "fixed",
  forecast_dates,
  intervals,
  horizon,
  truth_source = "JHU",
  plot = TRUE,
  truth_as_of = NULL,
  title = "default",
  show_caption = TRUE
)
}
\arguments{
\item{forecast_data}{data frame with truth and forecasts from load_forecasts()}

\item{truth_data}{optional data frame with forecasts in the format returned
by load_truth().}

\item{models}{model_abbr specifying models to plot. Optional if there is only
one model available in forecast data.}

\item{target_variable}{string specifying target type. It should be one of
"cum death", "inc case", "inc death"}

\item{locations}{string for fips code or 'US'. Optional if there is only one
location available in forecast data.}

\item{facet}{interpretable facet option for ggplot}

\item{facet_scales}{argument for scales in ggplot2::facet_wrap. Default to "fixed".}

\item{forecast_dates}{date string vectors for forecast dates to plot.
Default to forecast_dates present in the data.}

\item{intervals}{values indicating which central prediction interval levels
to plot. NULL means only plotting point forecasts.
If not provided, it will default to c(.5, .8, .95).}

\item{horizon}{forecasts are plotted for the horizon time steps after the
forecast date. Default to all available horizons in forecast data.}

\item{truth_source}{character specifying where the truth data will
be loaded from if truth_data is not provided.
Otherwise, this character specifies the data source to plot.
Currently support "JHU","USAFacts" and "NYTimes".}

\item{plot}{boolean for showing the plot. Default to TRUE.
Currently supports "JHU","USAFacts", "NYTimes". Default to "JHU".}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the truth_as_of date (not using this parameter when truth data
is from github repo)}

\item{title}{optional text for the title of the plot. if left as "default",
the title will be automatically generated. if "none", no title will be plotted.}

\item{show_caption}{logical, if TRUE, caption will be included showing data sources}
}
\value{
invisible ggplot object
}
\description{
Plot forecasts with truth for only one model, one target and one location
}
