% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecasts.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecasts and optional truth for only one selected target variable.
Faceted plots for multiple models, locations and forecast dates are
supported with specified facet formula.}
\usage{
plot_forecast(
  forecast_data,
  truth_data = NULL,
  hub = hub,
  models,
  target_variable,
  locations,
  facet = NULL,
  facet_scales = "fixed",
  facet_nrow = NULL,
  facet_ncol = NULL,
  forecast_dates,
  intervals,
  horizon,
  truth_source,
  use_median_as_point = FALSE,
  plot_truth = TRUE,
  plot = TRUE,
  fill_by_model = FALSE,
  fill_transparency = 1,
  truth_as_of = NULL,
  title = "default",
  subtitle = "default",
  show_caption = TRUE
)
}
\arguments{
\item{forecast_data}{data frame with truth and forecasts from load_forecasts()}

\item{truth_data}{optional data frame from one truth source in the format returned
by load_truth(). It needs to have columns model, target_variable,
target_end_date, location and value.
Model column can be "Observed Data (a truth source)".}

\item{hub}{character, which hub to use.
Default is "US", other option is "ECDC".}

\item{models}{vector of strings specifying models to plot.
Default to all models in forecast_data.}

\item{target_variable}{string specifying target type. It should be one of
"cum death", "inc case", "inc death" and "inc hosp".
"cum death" and "inc hosp" are only available in forecasts from US hub now.
If there is only one target_variable in forecast_data, this parameter is optional.}

\item{locations}{string for fips code, 'US' or abbreviation of European countries.
Default to all locations in forecast_data.}

\item{facet}{interpretable facet option for ggplot. Function will error
if multiple locations are passed in without location in the facet formula.}

\item{facet_scales}{argument for scales in ggplot2::facet_wrap. Default to "fixed".}

\item{facet_nrow}{number of rows for facetting; optional.}

\item{facet_ncol}{number of columns for facetting; optional.}

\item{forecast_dates}{date string vectors for forecast dates to plot.
Default to forecast_dates present in the data.}

\item{intervals}{values indicating which central prediction interval levels
to plot. NULL means only plotting point forecasts.
If not provided, it will default to c(.5, .8, .95).
When plotting 6 models or more, the plot will be reduced to show .95 interval only.}

\item{horizon}{forecasts are plotted for the horizon time steps after the
forecast date. Default to all available horizons in forecast data.}

\item{truth_source}{character specifying where the truth data will
be loaded from if truth_data is not provided. Currently support "JHU",
"USAFacts", "NYTimes", "HealthData" and "ECDC".
Optional if truth_data is provided.}

\item{use_median_as_point}{boolean for using median quantiles as point forecasts in plot.
Default to FALSE.}

\item{plot_truth}{boolean for showing truth data in plot. Default to FALSE.}

\item{plot}{boolean for showing the plot. Default to TRUE.
Currently supports "JHU","USAFacts", "NYTimes". Default to "JHU".}

\item{fill_by_model}{boolean for specifying colors in plot.
If TRUE, separate colors will be used for each model.
If FALSE, only blues will be used for all models. Default to FALSE.}

\item{fill_transparency}{numeric value used to set transparency of intervals.
0 means fully transparent, 1 means opaque.}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the truth_as_of date (not using this parameter when truth data
is from github repo)}

\item{title}{optional text for the title of the plot. If left as "default",
the title will be automatically generated. If "none", no title will be plotted.}

\item{subtitle}{optional text for the subtitle of the plot. If left as "default",
the subtitle will be automatically generated. If "none", no subtitle will be plotted.}

\item{show_caption}{logical, if TRUE, caption will be included showing data sources}
}
\value{
invisible ggplot object
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}
Please use \code{\link[=plot_forecasts]{plot_forecasts()}} instead.
}
