% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecasts with truth for only one model, one target and one location}
\usage{
plot_forecast(
  forecast_data,
  truth_data = NULL,
  model,
  target_variable,
  location,
  intervals = c(0.5, 0.8, 0.95),
  horizon,
  truth_source = "JHU",
  plot = TRUE,
  truth_as_of = NULL
)
}
\arguments{
\item{forecast_data}{data frame with truth and forecasts from load_forecasts()}

\item{truth_data}{optional data frame with forecasts in the format returned
by load_truth().}

\item{model}{model_abbr specifying model to plot}

\item{target_variable}{string specifying target type. It should be one of
"cum death", "inc case", "inc death"}

\item{location}{string for fips code or 'US'}

\item{intervals}{values indicating which central prediction interval levels
to plot, defaults to c(.5, .8, .95). NULL means only plotting point forecasts.}

\item{horizon}{forecasts are plotted for the horizon time steps after the
forecast date}

\item{truth_source}{character specifying where the truth data will
be loaded from if truth_data is not provided.
Otherwise, this character specifies the data source to plot.
Currently support "JHU","USAFacts" and "NYTimes".}

\item{plot}{boolean for showing the plot. Default to TRUE.
Currently supports "JHU","USAFacts", "NYTimes". Default to "JHU".}

\item{truth_as_of}{the plot includes the truth data that would have been
in real time as of the truth_as_of date (not using this parameter when truth data
is from github repo)}
}
\value{
invisible ggplot object
}
\description{
Plot forecasts with truth for only one model, one target and one location
}
