% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_jhu}
\alias{preprocess_jhu}
\title{Preprocess historical raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases and write to CSVs.
Note: updates to the JHU CSSE data stopped in March 2023.}
\usage{
preprocess_jhu(save_location = "./data-truth")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to \code{"./data-truth"}}
}
\description{
Preprocess historical raw truth data from JHU CSSE into Cumulative/Incident - Deaths/Cases and write to CSVs.
Note: updates to the JHU CSSE data stopped in March 2023.
}
