% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_truth.R
\name{preprocess_nytimes}
\alias{preprocess_nytimes}
\title{Preprocess raw truth data from NYTimes into Cumulative/Incident - Deaths/Cases and write to CSVs}
\usage{
preprocess_nytimes(save_location = "./data-truth/nytimes/")
}
\arguments{
\item{save_location}{character specifying the location of to save raw truth data.
Default to "./data-truth/nytimes/raw/"}
}
\description{
Preprocess raw truth data from NYTimes into Cumulative/Incident - Deaths/Cases and write to CSVs
}
