% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_forecasts_zoltar.R
\name{reformat_forecasts}
\alias{reformat_forecasts}
\title{Reformat forecast data frame returned from zoltar query
This function will throw a warning and return an empty data frame with
columns model, forecast_date, location, type, quantile and value, when
no forecasts are available in zoltar_query_result.}
\usage{
reformat_forecasts(zoltar_query_result)
}
\arguments{
\item{zoltar_query_result}{dataframe returned by zoltr::do_zoltar_query}
}
\value{
data frame with columns model, forecast_date, location, horizon,
temporal_resolution, target_variable, target_end_date, type, quantile, value
}
\description{
Reformat forecast data frame returned from zoltar query
This function will throw a warning and return an empty data frame with
columns model, forecast_date, location, type, quantile and value, when
no forecasts are available in zoltar_query_result.
}
